Abstract prometheus implementation from metrics plugin
