+++ plugins/metrics/plugin.go
@@ -1,23 +1,19 @@
 package metrics
 
 import (
-	"context"
-	"net/http"
-	"time"
-
 	"github.com/iotaledger/hive.go/daemon"
 	"github.com/iotaledger/hive.go/logger"
 	"github.com/iotaledger/hive.go/node"
+	"github.com/iotaledger/wasp/packages/metrics"
 	"github.com/iotaledger/wasp/packages/parameters"
-	"github.com/labstack/echo/v4"
-	"github.com/labstack/echo/v4/middleware"
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
 
 const PluginName = "Metrics"
 
-var log *logger.Logger
+var (
+	log            *logger.Logger
+	metricsManager *metrics.Metrics
+)
 
 func Init() *node.Plugin {
 	return node.NewPlugin(PluginName, node.Disabled, configure, run)
@@ -25,6 +21,7 @@ func Init() *node.Plugin {
 
 func configure(_ *node.Plugin) {
 	log = logger.NewLogger(PluginName)
+	metricsManager = metrics.New()
 }
 
 func run(_ *node.Plugin) {
@@ -36,30 +33,11 @@ func run(_ *node.Plugin) {
 	if err := daemon.BackgroundWorker("Prometheus exporter", func(shutdownSignal <-chan struct{}) {
 		log.Info("Starting Prometheus exporter ... done")
 
-		e := echo.New()
-		e.HideBanner = true
-		e.Use(middleware.Recover())
-
-		e.GET("/metrics", func(c echo.Context) error {
-			handler := promhttp.HandlerFor(
-				prometheus.DefaultGatherer,
-				promhttp.HandlerOpts{
-					EnableOpenMetrics: true,
-				},
-			)
-			handler.ServeHTTP(c.Response(), c.Request())
-			return nil
-		})
-
-		bindAddr := parameters.GetString(parameters.MetricsBindAddress)
-		server := &http.Server{Addr: bindAddr, Handler: e}
-
 		stopped := make(chan struct{})
 		go func() {
 			defer close(stopped)
-			log.Infof("%s started, bind-address=%s", PluginName, bindAddr)
-			if err := server.ListenAndServe(); err != nil {
-				log.Warn("Error serving: %s", err)
+			if err := metricsManager.Start(); err != nil {
+				log.Warnf("Error serving: %s", err)
 			}
 		}()
 		select {
@@ -68,9 +46,7 @@ func run(_ *node.Plugin) {
 		}
 		log.Info("Stopping %s ...", PluginName)
 		defer log.Infof("Stopping %s ... done", PluginName)
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-		if err := server.Shutdown(ctx); err != nil {
+		if err := metricsManager.Stop(); err != nil {
 			log.Errorf("Error stopping: %s", err)
 		}
 	}, parameters.PriorityMetrics); err != nil {
